requirements.txt
fastapi
uvicorn
python-dotenv
openai
PyPDF2
markitdown[all]
pandas
python-multipart
aiofiles